sms-management-system/backend/app/requirements.txt
fastapi
uvicorn
pymongo
mysql-connector-python
prometheus-client
python-jose
requests
